+++ examples/discord-bots/moderation/index.js
@@ -1,6 +1,7 @@
 const { defineBot } = require("discord");
 const registerOverviewCommands = require("./lib/register-overview-commands");
 const registerMessageModerationCommands = require("./lib/register-message-moderation-commands");
+const registerChannelModerationCommands = require("./lib/register-channel-moderation-commands");
 const registerMemberModerationCommands = require("./lib/register-member-moderation-commands");
 const registerModerationEvents = require("./lib/register-events");
 
@@ -13,6 +14,7 @@ module.exports = defineBot((api) => {
 
   registerOverviewCommands(api);
   registerMessageModerationCommands(api);
+  registerChannelModerationCommands(api);
   registerMemberModerationCommands(api);
   registerModerationEvents(api);
 });